+++ authentication-server/src/main/java/com/wesell/authenticationserver/controller/AuthController.java
@@ -1,9 +1,9 @@
 package com.wesell.authenticationserver.controller;
 
-import com.wesell.authenticationserver.dto.GeneratedTokenDto;
-import com.wesell.authenticationserver.dto.request.CreateUserRequestDto;
-import com.wesell.authenticationserver.dto.request.SignInUserRequestDto;
-import com.wesell.authenticationserver.dto.response.SignInSuccessResponseDto;
+import com.wesell.authenticationserver.controller.dto.GeneratedTokenDto;
+import com.wesell.authenticationserver.controller.dto.request.CreateUserRequestDto;
+import com.wesell.authenticationserver.controller.dto.request.SignInUserRequestDto;
+import com.wesell.authenticationserver.service.dto.response.SignInSuccessResponseDto;
 import com.wesell.authenticationserver.global.util.CustomCookie;
 import com.wesell.authenticationserver.response.CustomException;
 import com.wesell.authenticationserver.response.ErrorCode;
