+++ authentication-server/src/main/java/com/wesell/authenticationserver/controller/dto/GeneratedTokenDto.java
@@ -1,4 +1,4 @@
-package com.wesell.authenticationserver.dto;
+package com.wesell.authenticationserver.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
