+++ authentication-server/src/main/java/com/wesell/authenticationserver/controller/dto/request/AdminAuthIsForcedRequestDto.java
@@ -1,6 +1,5 @@
-package com.wesell.authenticationserver.dto.request;
+package com.wesell.authenticationserver.controller.dto.request;
 
-import com.wesell.authenticationserver.domain.enum_.Role;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +9,6 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AdminAuthRoleRequestDto {
+public class AdminAuthIsForcedRequestDto {
     private String uuid;
-    private Role role;
 }
