+++ authentication-server/src/main/java/com/wesell/authenticationserver/controller/dto/request/SignInUserRequestDto.java
@@ -1,4 +1,4 @@
-package com.wesell.authenticationserver.dto.request;
+package com.wesell.authenticationserver.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
