--- authentication-server/src/main/java/com/wesell/authenticationserver/dto/response/AdminAuthIsForcedResponseDto.java
@@ -1,14 +1,0 @@
-package com.wesell.authenticationserver.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AdminAuthIsForcedResponseDto {
-    private String message;
-}
