+++ authentication-server/src/main/java/com/wesell/authenticationserver/global/util/CustomConverter.java
@@ -2,9 +2,9 @@ package com.wesell.authenticationserver.global.util;
 
 import com.wesell.authenticationserver.domain.entity.AuthUser;
 import com.wesell.authenticationserver.domain.enum_.Role;
-import com.wesell.authenticationserver.dto.feign.AuthUserListFeignResponseDto;
-import com.wesell.authenticationserver.dto.request.CreateUserRequestDto;
-import com.wesell.authenticationserver.dto.response.CreateUserFeignResponseDto;
+import com.wesell.authenticationserver.service.dto.feign.AuthUserListFeignResponseDto;
+import com.wesell.authenticationserver.controller.dto.request.CreateUserRequestDto;
+import com.wesell.authenticationserver.service.dto.response.CreateUserFeignResponseDto;
 import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -47,7 +47,7 @@ public class CustomConverter {
     }
 
     public List<AuthUserListFeignResponseDto> toFeignDto(List<AuthUser> authUserList){
-        return authUserList.stream().map(
+        return authUserList.stream().filter(auth->!auth.isDelete()).map(
                 user->AuthUserListFeignResponseDto.builder()
                         .uuid(user.getUuid())
                         .email(user.getEmail())
