+++ authentication-server/src/main/java/com/wesell/authenticationserver/service/AdminAuthService.java
@@ -3,8 +3,8 @@ package com.wesell.authenticationserver.service;
 import com.wesell.authenticationserver.domain.entity.AuthUser;
 import com.wesell.authenticationserver.domain.enum_.Role;
 import com.wesell.authenticationserver.domain.repository.AdminAuthUserRepository;
-import com.wesell.authenticationserver.dto.request.AdminAuthIsForcedRequestDto;
-import com.wesell.authenticationserver.dto.response.AdminAuthIsForcedResponseDto;
+import com.wesell.authenticationserver.controller.dto.request.AdminAuthIsForcedRequestDto;
+import com.wesell.authenticationserver.service.dto.response.AdminAuthIsForcedResponseDto;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
