+++ authentication-server/src/main/java/com/wesell/authenticationserver/service/dto/response/CreateUserFeignResponseDto.java
@@ -1,4 +1,4 @@
-package com.wesell.authenticationserver.dto.response;
+package com.wesell.authenticationserver.service.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
